Create minimum_operations.py
Given a array you are allowed to do the following operation on it
add either 1 or -1 to a subarray a[ i : ] where i can be anything. The value will be added to every element of the subarray. 
print the minimum number of such operations where one operation is defined to add once